+++ types/data_file.go
@@ -10,6 +10,12 @@ type DataFile interface {
 	HasAnyTargetedDeliveryRule() bool
 	GetFeatureFlagById(featureFlagId int) FeatureFlag
 	GetRuleBySegmentId(segmentId int) Rule
+	GetRuleInfoByExpId(experimentId int) (RuleInfo, bool)
 	GetVariation(variationId int) *VariationByExposition
 	HasExperimentJsCssVariable(experimentId int) bool
+}
+
+type RuleInfo struct {
+	FeatureFlag FeatureFlag
+	Rule        Rule
 }