+++ types/rule_base.go
@@ -1,5 +1,7 @@
 package types
 
+import "github.com/Kameleoon/client-go/v3/errs"
+
 type RuleType uint8
 
 const (
@@ -35,4 +37,32 @@ type RuleBase struct {
 	ExperimentId          int                     `json:"experimentId,omitempty"`
 	VariationByExposition []VariationByExposition `json:"variationByExposition"`
 	RespoolTime           int                     `json:"respoolTime,omitempty"`
+}
+
+func (r *RuleBase) GetVariationByHash(hashDouble float64) *VariationByExposition {
+	threshold := hashDouble
+	for _, varByExp := range r.VariationByExposition {
+		threshold -= varByExp.Exposition
+		if threshold < 0 {
+			return &varByExp
+		}
+	}
+	return nil
+}
+
+func (r *RuleBase) GetVariationByKey(variationKey string) (*VariationByExposition, error) {
+	for i := range r.VariationByExposition {
+		if r.VariationByExposition[i].VariationKey == variationKey {
+			return &r.VariationByExposition[i], nil
+		}
+	}
+	return nil, errs.NewFeatureVariationNotFoundWithVariationKey(r.Id, variationKey)
+}
+
+func (r *RuleBase) IsExperimentType() bool {
+	return r.Type == RuleTypeExperimentation
+}
+
+func (r *RuleBase) IsTargetDeliveryType() bool {
+	return r.Type == RuleTypeTargetedDelivery
 }