+++ utils/nonce.go
@@ -3,5 +3,5 @@ package utils
 const NonceLength = 16
 
 func GetNonce() string {
-	return GetRandomString(NonceLength)
+	return GetRandomString(NonceLength, "ABCDEF0123456789")
 }