+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 3.7.0 - 2024-12-16
+### Features
+* Added support for **simulated** variations.
+* Added the [`SetForcedVariation()`](https://developers.kameleoon.com/feature-management-and-experimentation/web-sdks/go-sdk#setforcedvariation) method. This method allows explicitly setting a forced variation for a visitor, which will be applied during experiment evaluation.
+### Bug fixes
+* Fixed the [`Variation.IsActive()`](https://developers.kameleoon.com/feature-management-and-experimentation/web-sdks/go-sdk#variation) method, which was returning an incorrect value.
+* Fixed an issue where the return values in the trace logs for the following methods of the `kameleoonClient` were always default:
+  - `IsFeatureActive`
+  - `IsFeatureActiveWithTracking`
+  - `isFeatureActive`
+  - `GetVariation`
+  - `GetVariations`
+  - `makeExternalVariation`
+
 ## 3.6.1 - 2024-11-20
 ### Bug fixes
 * Resolved an issue where the validation of [top-level domains](https://developers.kameleoon.com/feature-management-and-experimentation/web-sdks/go-sdk/#additional-configuration) for `localhost` resulted in incorrect failures. The SDK now accepts the provided domain without modification if it is deemed invalid and logs an [error](https://developers.kameleoon.com/feature-management-and-experimentation/web-sdks/go-sdk/#log-levels) to notify you of any issues with the specified domain.